chore: move product related types
