+++ src/Domain/Purchase/ProductPurchase.php
@@ -16,11 +16,11 @@ final readonly class ProductPurchase
     public function __construct(
         public string $kind,
         public GooglePlay\Time $purchaseTimeMillis,
-        public GooglePlay\PurchaseState $purchaseState,
-        public GooglePlay\ConsumptionState $consumptionState,
+        public PurchaseState $purchaseState,
+        public ConsumptionState $consumptionState,
         public string $orderId,
-        public GooglePlay\PurchaseType $purchaseType,
-        public GooglePlay\ProductAcknowledgementState $acknowledgementState,
+        public PurchaseType $purchaseType,
+        public ProductAcknowledgementState $acknowledgementState,
         public string $purchaseToken,
         public string $productId,
         public string $regionCode,
